+++ apps/dokploy/docker/build.sh
@@ -12,6 +12,8 @@ fi
 
 BUILDER=$(docker buildx create --use)
 
-docker buildx build --platform linux/amd64,linux/arm64 --pull --rm -t "dokploy/dokploy:${TAG}" -f 'Dockerfile' .
+DOKPLOY_IMAGE_REPO="${DOKPLOY_IMAGE_REPO:-ghcr.io/rain-kl/dokploy}"
+
+docker buildx build --platform linux/amd64,linux/arm64 --pull --rm -t "${DOKPLOY_IMAGE_REPO}:${TAG}" -f 'Dockerfile' .
 
 docker buildx rm $BUILDER